// src/interfaces/AcceptServiceInterface.tsx
// Generated by https://quicktype.io

export interface getCollectionParams {
  rowsPerPage: number,
  fechaProgInicial?: string,
  fechaProgFinal?: string,
  reload?: boolean,
}

export interface updateParams {
  id: number,
  observaciones_rechazo: string,
  indicativo_aceptacion: string,
  tipo_proceso: string,
  accion: string
}

export interface OrdenServicio {
  datos:         Dato[];
  desde:         number;
  hasta:         number;
  por_pagina:    number;
  pagina_actual: number;
  ultima_pagina: number;
  total:         number;
}

export interface Dato {
  id:                         number;
  numero_orden_servicio:      number;
  fecha_creacion:             string;
  estado_orden_servicio:      string;
  asociado:                   string;
  fecha_programada:           string;
  hora_programada:            string;
  departamento:               string;
  departamento_id:            number;
  ciudad:                     string;
  lugar:                      string;
  direccion:                  string;
  recurso:                    null | string;
  tipo_servicio:              string;
  numero_viaje:               number | null;
  recurso_id:                 number | null;
  equipo_id:                  number;
  numero_serial:              string;
  equipo:                     string;
  observaciones_programacion: null | string;
  indicativo_aceptacion:      null | string;
  observaciones_rechazo:      null | string;
}